+++ src/an-common/bl.h
@@ -142,7 +142,7 @@ Inline int  il_get(il* list, const int n);
 void il_copy(il* list, int start, int length, int* vdest);
 il*  il_dupe(il* list);
 void il_print(il* list);
-void   il_insert(il* list, int indx, double data);
+void   il_insert(il* list, int indx, int data);
 int il_insert_ascending(il* list, int n);
 int il_insert_descending(il* list, int n);
 int  il_insert_unique_ascending(il* list, int p);
